Explicitly disable unused utilities
Add configure options to explicitly disable the utilities that
we do not ship because they would conflict with what other packages
already ship.

These should all be reviewed together with the other packages maintainer
because maybe we should start using the util-linux implemenation instead.

This just cuts down on "list-missing spam" for now.