Use correct configure options on non-linux
The enable raw and enable tunelp is only valid on Linux, so move
them to the linux-only section.

The pivot_root option only seems to work on Linux as well, so
explicitly disable it on non-linux to avoid configure check failing.

While at it, also split long multi-options lines into multiple lines.