Skip installing a whole bunch of utils on non-linux
The difference from util-linux 2.20.1-5.8 on kFreeBSD is
that fsck, readprofile, linux32 and linux64 are no longer shipped.

linux* was only dead symlinks and readprofile is supposedly
also linux-specific, so these where a mistake to ship earlier.

The problem with fsck is that it now relies on libmount, which
needs porting to non-linux.
See upstream commit 2b5051242ec082f12126ae552d12d79d590fe86f
(and a few after, which leverage the new libmount integration).