Fine-tune hwclock.sh init script LSB dependencies
In /etc/insserv.conf, $local_fs is defined as

   $local_fs       +mountall +mountall-bootclean +mountoverflowtmp +umountfs

hwclock.sh starts before checkroot.sh (fsck needs a clock) which in turn
is needed by mountall.sh (fsck before mount).

If hwclock.sh depends on $local_fs (needs to save /etc/adjtime before
remount / ro), there is a dependency loop.

At present, the problem is hided by giving a partial dependency of
Required-Stop, but not Required-Start.

It's very nice if Required-Stop/Start pair is treated on the same foot,
especially for OpenRC, which does not differentiate between the two.

As hwclock.sh only need to be stopped before umountfs, not the other
services in $local_fs, it is reasonable to replace "Required-Stop:
$local_fs" with "Should-Stop: umountfs". ($local_fs optionally
depends on +umountfs, therefore Should-Stop)

Closes: #736400