Fix sparc install mangling
The util-linux.install and util-linux.manpages files are overridden
by (temporary) *.linux files, so it's those that need to be mangled.

While at it make the mangled files explicit and create a
temporary .linux install file for fdisk-udeb as well...
This means everything now gets cleaned up properly on sparc.