Use dh-exec (>= 0.13)
With the new features in the brand new version of dh-exec, in particular
the arch filtering feature, the util-linux package can be further
simplified / cleaned up.

Some changes in here actually only use already existing dh_install
features (eg. for moving files to different directory) and are not
dependent on dh-exec.

Also note that the install of bash completions was changed to use
mv instead of dh_install, as dh_install even when specifying
what to install on the command line would (re)install all of the
things mentioned in debian/$(package).install ....

Along the lines a bug was noticed on non-linux. The /sbin/mkswap
(along with its manpage) was supposed to be renamed /sbin/mkswap.linux
but *both* of them where installed in the package.

This bug is potentially fixed in here along with /sbin/getty being a
proper symlink to agetty instead of a copy, otherwise this should have
no functional changes on the resulting packages.

(Related note: /sbin/getty is shipped by "hurd" package on Hurd.)